+++ documentation/playground/src/Playground.tsx
@@ -34,6 +34,7 @@ import ListExamples from './atoms/ListExamples';
 import RangeSliderExamples from './atoms/RangeSliderExamples';
 import AutocompleteExamples from './molecules/AutocompleteExamples';
 import DatepickerExamples from './molecules/DatepickerExamples';
+import SelectExamples from './atoms/SelectExamples';
 
 export function Playground() {
   return (
@@ -51,6 +52,7 @@ export function Playground() {
       <ProgressBarExamples />
       <RadioExamples />
       <RangeSliderExamples />
+      <SelectExamples />
       <SpinnerExamples />
       <SwitchExamples />
       <TableExamples />